tests: consolidate rows test cases
Refs #234
