--- tests/AI Cabinets/TC_Rows_Reveal_OverlayTypes.rb
@@ -1,49 +1,0 @@
-# frozen_string_literal: true
-
-require 'testup/testcase'
-
-require_relative 'suite_helper'
-require_relative '../support/model_query'
-require_relative '../support/rows_reveal_helpers'
-require_relative '../support/rows_test_harness'
-
-Sketchup.require('aicabinets/rows')
-Sketchup.require('aicabinets/rows/reveal')
-Sketchup.require('aicabinets/test_harness')
-
-class TC_Rows_Reveal_OverlayTypes < TestUp::TestCase
-  include RowsRevealTestHelpers
-
-  REVEAL_MM = 3.0
-  TOLERANCE_MM = 0.1
-
-  def setup
-    RowsTestHarness.reset_model!
-  end
-
-  def teardown
-    RowsTestHarness.reset_model!
-  end
-
-  def test_reveal_applies_to_supported_overlay_types
-    [:frameless_overlay, :face_frame_overlay].each do |overlay_type|
-      row_id, instances = build_row(widths_mm: [600.0, 500.0, 500.0], overlay_type: overlay_type)
-      apply_row_reveal!(row_id: row_id, reveal_mm: REVEAL_MM)
-
-      first, second, third = instances
-      gap_left_middle = interior_gap_mm(first, second)
-      gap_middle_right = interior_gap_mm(second, third)
-
-      AICabinetsTestHelper.assert_within_tolerance(self, REVEAL_MM, gap_left_middle, TOLERANCE_MM)
-      AICabinetsTestHelper.assert_within_tolerance(self, REVEAL_MM, gap_middle_right, TOLERANCE_MM)
-
-      left_end = left_end_gap_mm(first)
-      right_end = right_end_gap_mm(third)
-
-      AICabinetsTestHelper.assert_within_tolerance(self, REVEAL_MM, left_end, TOLERANCE_MM)
-      AICabinetsTestHelper.assert_within_tolerance(self, REVEAL_MM, right_end, TOLERANCE_MM)
-
-      RowsTestHarness.reset_model!
-    end
-  end
-end
